perf(core): SIMD + uint8 pixel path single-thread WASM decode speedups

+++ src/decoding/coding_tree.h
@@ -79,8 +79,10 @@ struct DecodingContext {
     SaoParams* sao_params = nullptr;   // array [PicWidthInCtbsY * PicHeightInCtbsY]
     int sao_params_stride = 0;         // = PicWidthInCtbsY
 
-    // Phase 9: SAO backup planes (reused across frames, owned by Decoder)
-    std::vector<uint16_t>* sao_backup = nullptr;  // array of 3 vectors
+    // Phase 9: SAO backup planes (reused across frames, owned by Decoder).
+    // Byte-sized: holds bytes_per_sample bytes/sample (uint8 native 8-bit or
+    // uint16); apply_sao reinterprets it as the plane's Sample type.
+    std::vector<uint8_t>* sao_backup = nullptr;  // array of 3 byte-vectors
 
     // Phase 10: slice index per CTU (for cross-slice boundary detection)
     uint8_t* slice_idx = nullptr;          // array [PicSizeInCtbsY], indexed by CtbAddrInRs
